+++ tests/entities-change.test.ts
@@ -1,7 +1,6 @@
 import HomeAssistantJavaScriptTemplates, { HomeAssistantJavaScriptTemplatesRenderer } from '../src';
 import { SubscriberEvent, HomeAssistant } from '../src/types';
-import { HOME_ASSISTANT_ELEMENT, HASS } from './constants';
-import { SUBSCRIBE_EVENTS, STATE_CHANGE_EVENT } from '../src/constants';
+import { HASS } from './constants';
 
 const CUSTOM_EVENT = 'subscribe_events';
 const getSubscribeCustomEvent = (id: string) => {
@@ -47,6 +46,19 @@ describe('promise instance', () => {
         expect(renderingFunction).toHaveBeenCalledWith('off');
     });
 
+    it('tracking the same template with multiple fucntions should call all of them', async () => {
+        const renderingFunction1 = jest.fn();
+        const renderingFunction2 = jest.fn();
+        const renderingFunction3 = jest.fn();
+        const template = 'states["light.woonkamer_lamp"].state';
+        renderer.trackTemplate(template, renderingFunction1);
+        renderer.trackTemplate(template, renderingFunction2);
+        renderer.trackTemplate(template, renderingFunction3);
+        expect(renderingFunction1).toHaveBeenCalledWith('off');
+        expect(renderingFunction2).toHaveBeenCalledWith('off');
+        expect(renderingFunction3).toHaveBeenCalledWith('off');
+    });
+
     it('tracking a template should call an update of the renderingFunction when an entity changes', async () => {
         const renderingFunction = jest.fn();
         renderer.trackTemplate('states["light.woonkamer_lamp"].state', renderingFunction);
@@ -57,6 +69,29 @@ describe('promise instance', () => {
         );
         expect(renderingFunction).toHaveBeenNthCalledWith(2, 'on');
         expect(renderingFunction).not.toHaveBeenCalledTimes(3);
+    });
+
+    it('tracking the same template with multiple functions should call an update of all the renderingFunctions when an entity changes', async () => {
+        const renderingFunction1 = jest.fn();
+        const renderingFunction2 = jest.fn();
+        const renderingFunction3 = jest.fn();
+        const template = 'states["light.woonkamer_lamp"].state';
+        renderer.trackTemplate(template, renderingFunction1);
+        renderer.trackTemplate(template, renderingFunction2);
+        renderer.trackTemplate(template, renderingFunction3);
+        expect(renderingFunction1).toHaveBeenNthCalledWith(1, 'off');
+        expect(renderingFunction2).toHaveBeenNthCalledWith(1, 'off');
+        expect(renderingFunction3).toHaveBeenNthCalledWith(1, 'off');
+        hassClone.states['light.woonkamer_lamp'].state = 'on';
+        window.dispatchEvent(
+            getSubscribeCustomEvent('light.woonkamer_lamp')
+        );
+        expect(renderingFunction1).toHaveBeenNthCalledWith(2, 'on');
+        expect(renderingFunction1).not.toHaveBeenCalledTimes(3);
+        expect(renderingFunction2).toHaveBeenNthCalledWith(2, 'on');
+        expect(renderingFunction2).not.toHaveBeenCalledTimes(3);
+        expect(renderingFunction3).toHaveBeenNthCalledWith(2, 'on');
+        expect(renderingFunction3).not.toHaveBeenCalledTimes(3);
     });
 
     it('tracking a template with an unreached entity should not call renderingFunction if that entity changes', async () => {
@@ -108,6 +143,40 @@ describe('promise instance', () => {
         );
         expect(renderingFunction1).toHaveBeenNthCalledWith(2, 'yes');
         expect(renderingFunction2).toHaveBeenNthCalledWith(2, 'on');
+    });
+
+    it('if the untrack function that trackTemplate returns is called then it should clean all the trackings for that template/rendering function', () => {
+        const renderingFunction1 = jest.fn();
+        const renderingFunction2 = jest.fn();
+        const untrack1 = renderer.trackTemplate(
+            `
+                if (is_state('light.woonkamer_lamp', 'on')) {
+                    return 'yes';
+                }
+                return 'no';
+            `,
+            renderingFunction1
+        );
+        const untrack2 = renderer.trackTemplate(
+            'states("light.woonkamer_lamp")',
+            renderingFunction2
+        );
+        expect(renderingFunction1).toHaveBeenNthCalledWith(1, 'no');
+        expect(renderingFunction2).toHaveBeenNthCalledWith(1, 'off');
+        untrack1();
+        hassClone.states['light.woonkamer_lamp'].state = 'on';
+        window.dispatchEvent(
+            getSubscribeCustomEvent('light.woonkamer_lamp')
+        );
+        expect(renderingFunction1).toHaveBeenCalledTimes(1);
+        expect(renderingFunction2).toHaveBeenNthCalledWith(2, 'on');
+        untrack2();
+        hassClone.states['light.woonkamer_lamp'].state = 'off';
+        window.dispatchEvent(
+            getSubscribeCustomEvent('light.woonkamer_lamp')
+        );
+        expect(renderingFunction1).toHaveBeenCalledTimes(1);
+        expect(renderingFunction2).toHaveBeenCalledTimes(2);
     });
 
     it('after executing cleanTracked the rendering functions should not be called anymore', async () => {